backend/pyproject.toml
[project]
name = "foodcalorie-backend"
version = "0.1.0"
description = "FoodCalorie Backend Django Application"
requires-python = ">=3.10"
dependencies = [
    "celery>=5.5.3",
    "channels==3.0.5",
    "channels-redis>=4.2.1",
    "django>=5.2.4",
    "pandas>=2.3.1",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "websockets>=15.0.1",
    "djangorestframework>=3.16.0",
    "django-cors-headers>=4.4.0",
    "Pillow>=10.3.0",
    "djangorestframework-simplejwt>=5.0.0",
    "django-filter>=23.0.0",
    "faker>=18.0.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "django-extensions>=3.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.flake8]
max-line-length = 79
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    ".venv",
    "venv",
    "env"
]